Extract sun body silhouette module
